Add GridItemSelect, now with previews
