+++ src/components/index.ts
@@ -1,9 +1,9 @@
 export * from './icons';
 export * from './MotionStatus';
 export * from './MotionSummary';
-export * from './BatteryLevel';
-export * from './BatterySummary';
+export * from './items/BatteryLevel';
 export * from './Card';
 export * from './Toggle';
 export * from './Sidebar';
-export * from './Camera';
+export * from './GridItemSelect';
+export * from './items';
